+++ @types/helpers/constants.d.ts
@@ -1,5 +1,6 @@
 export declare const mainConst: {
   defaultPriority: string[];
+  scssImportRegex: RegExp;
   sassImportRegex: RegExp;
   parseFilePathRegex: RegExp;
   fileType: string;