+++ @types/helpers/fs-utils.d.ts
@@ -3,4 +3,8 @@ export declare const readSync: (filePath: string) => string;
 export declare const cwDir: () => string;
 export declare const resolveDirDest: (fileDest: string) => void;
 export declare const writeAsync: (path: string, content: string) => Promise<{}>;
-export declare const defineExtension: (filePath: string) => string;
+export declare const fileType: (fileName: string) => string;
+export declare const defineExtension: (
+  filePath: string,
+  fileExtension: string
+) => string;