+++ @types/helpers/logger.d.ts
@@ -1,1 +1,1 @@
-export declare const logger: (info: string | Error) => void;
+export declare const logger: (info: string | Error) => any;