+++ src/helpers/logger.ts
@@ -1,1 +1,3 @@
-export const logger = (info: Error | string) => console.log(info);
+const log = require('fancy-log');
+
+export const logger = (info: Error | string) => log.info(info);